requirements.txt
tweepy
flask
flask-cors
numpy
pandas
python-dotenv
geopy